Add Dockerfile and update dependencies
Added a new Dockerfile for deploying the app with new environment variables. Also updated requirements.txt to include new packages needed for the Docker environment and additional functionality in the app. The Dockerfile enables us to build a Docker image for the application, ensuring it runs in an isolated and consistent environment.